Add standard tag generation to release of pypi package and add release trigger (#49)
## Version bump 
#minor: add tag to release of pypi package workflow

## Summary
It adds a tag to release workflow of a pypi package. Also the trigger
for a release is now dependent upon creating a release. Which is again
done by the release notes generator on the main branch

### Jira ticket
None, but forgot to update the version
[here](https://github.com/repowerednl/daphne/actions/runs/15729894379/job/44328439760)
and thought; a well. A 15min fix with testing

## Checks
- [x] I have tested this