minimal change for canonical image (#61)
## Version bump 
#patch

## Summary
Minimal implementation to add the missing 'docker.io/' part for the
image canonicalisation

### Questions
Q: do we want to separate the 'latest' image on a different output
variable (e.g. image_tag ; image_tag_latest)?

Q: when researching about this the idea of using a digest instead of
tags has popped up several times. Do we want to use digests instead of
tags for the platform?

### Jira ticket
[IN-283]



[IN-283]:
https://repowerednl.atlassian.net/browse/IN-283?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ