UV versions of existing workflows (#63)
## Version bump 
#patch

## Summary
I am not sure if this is necessary, or if this has already been made but
since we want to switch to UV for other packages as well, this would
need an update in the workflows as well?
I came across this when moving from poetry to uv in pixie - and saw that
the workflows are failing.
This is a setup after looking at the `pytest-uv-test-coverage.yml`, and
some AI help.

IT WORKS! See here:
https://github.com/repowerednl/repower-mlops/actions/runs/19037240462/job/54364686696

---------

Co-authored-by: Lucas Velvis <lucasvelvis@gmail.com>
Co-authored-by: LucasVelvis <60436680+LucasVelvis@users.noreply.github.com>
Co-authored-by: Ander Eguiluz <ander.eguiluz@repowered.nl>