In 64 prod target sync workflow (#66)
This workflow (+ template) enables a repository to synchronize ORM
postgres databases with eachother.

Triggers for the workflow can be:
- on a schedule; default is every Sunday at 18:00 UTC
- after successful release
- manually

Optionally, you can use GitHub environments to make sure that the
correct databases are synced w.r.t. branch it's run in.

'Succesful' workflow:
https://github.com/repowerednl/daphne/actions/runs/19103857965
I wanted to test the workflow here and not actually do a sync. So I
didn't configure it for a read database, hence the internal connection
error.